+++ Classes/Controller/UploadController.php
@@ -22,23 +22,29 @@ use TYPO3\CMS\Core\Page\JavaScriptModuleInstruction;
 use TYPO3\CMS\Core\Page\PageRenderer;
 use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
 use Wacon\Filetransfer\Domain\Model\Upload;
+use TYPO3\CMS\Core\Core\Environment;
+use Wacon\Filetransfer\Service\FileUploadService;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+use Wacon\Filetransfer\Exception\FileUploadException;
+use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
 
 final class UploadController extends ActionController
 {
+    protected string $extensionKey = 'filetransfer';
+
     public function __construct(private readonly PageRenderer $pageRenderer) {}
 
     /**
-     * Show the upoad form
-     * @return ResponseInterface
+     * Before any action function is called
+     * @return void
      */
-    public function formAction(): ResponseInterface
+    public function initializeAction()
     {
-        $this->pageRenderer->getJavaScriptRenderer()->addJavaScriptModuleInstruction(
-            JavaScriptModuleInstruction::create('@wacon/filetransfer/Fileupload.js')
-        );
-
-        $this->view->assign('contentObjectData', $this->request->getAttribute('currentContentObject')->data);
-        return $this->htmlResponse();
+        /* if ($this->request->hasAttribute('upload')) {
+            $this->arguments->getArgument('upload')
+                ->getPropertyMappingConfiguration()
+                ->setTargetTypeForSubProperty('validityDuration', 'int');
+        } */
     }
 
     /**
@@ -46,10 +52,60 @@ final class UploadController extends ActionController
      * @param Upload $upload
      * @return ResponseInterface
      */
-    public function uploadAction(Upload $upload): ResponseInterface
+    public function formAction(Upload $upload = null): ResponseInterface
     {
-        debug($upload);
+        if (!$upload) {
+            $upload = new Upload();
+        }
+
+        if ($this->request->hasArgument('testmode') && $this->request->getArgument('testmode') == '1') {
+            $upload->setTestData();
+        }
+
+        $this->pageRenderer->getJavaScriptRenderer()->addJavaScriptModuleInstruction(
+            JavaScriptModuleInstruction::create('@wacon/filetransfer/Fileupload.js')
+        );
+
+
+        $this->view->assign('upload', $upload);
+        $this->view->assign('contentObjectData', $this->request->getAttribute('currentContentObject')->data);
+        $this->view->assign('testMode', Environment::getContext()->isDevelopment());
         return $this->htmlResponse();
+    }
+
+    /**
+     * Show the upoad form
+     * @param Upload $upload
+     * @param array $asset
+     * @return ResponseInterface
+     * @throws FileUploadException
+     */
+    public function uploadAction(Upload $upload, array $asset): ResponseInterface
+    {
+        try {
+            $fileUploadService = GeneralUtility::makeInstance(FileUploadService::class);
+            $fileUploadService->init($this->settings['upload']);
+            $fileUploadService->uploadSingle($asset);
+            $asset = $fileUploadService->getFirstAsset();
+
+            if (!$asset) {
+                throw new FileUploadException(LocalizationUtility::translate('form.upload.fileuploadservice.upload_failed', $this->extensionKey));
+            }
+
+            $upload->setAsset($asset);
+        }catch(FileUploadException $e) {
+            $translatedMessage = LocalizationUtility::translate($e->getMessage(), $this->extensionKey);
+
+            // We need to do this, because I only set translation keys inside the FileUploadService
+            if ($translatedMessage) {
+                $e->setMessage($translatedMessage);
+            }
+
+            $this->view->assign('error', $e);
+            return $this->htmlResponse();
+        }
+
+        return $this->redirect('success');
     }
 
     /**